Fix up some discrepancies regarding environment variable passing
It seems like at one point the docker-compose.yml file's frontend container used the 'builder' image rather than the 'deps' image, passing in its environment variables as arguments to the image. However, the docker-compose.yml file has no way of overriding the builder image's "yarn build" command as the frontend Dockerfile suggests was intended, which is likely why the 'deps' image was used instead. This commit gets rid of the remnants of that iteration of our Docker Compose setup.